+++ CHANGELOG.md
@@ -20,6 +20,7 @@
  - preparing the environment may be optional
  - errors during test function are now saved into test record db
  - new config options (see readme)
+ - kill child processes in job timer
  
    
 ## [1.2.0]